avoid duplicate logic in the two classes called "TypingStrategyFilter"
